M2-1 完成: task 死亡撤销指向它的 cap (cap_id_t 32位回退)
M2-1 (保留): task_cleanup_resources 加 cap_revoke_object(),
撤销所有指向此 task 的 cap,防止 id 复用后陈旧授权。

M2-2 (回退): cap_id_t 16→32 位 + generation 扩展。
回退原因:syscall 测试 10 fail (cap transfer 相关),
32 位 cap_id_t 导致大量 BSS 数组翻倍 (TCB cap_set,
ep_server_reply_cap 等),虽然 BSS 未超 SRAM,
但 cap transfer 路径有未知截断/对齐问题。

保留 cap_decode/cap_get_entry 的 uint32_t 内部类型
(向后兼容,不影响 16 位 cap_id_t)。

generation 扩展留 M2 后续:需要先解决 cap transfer
对齐问题再扩位宽。

验证: Passed: 2878, All tests PASSED!

+++ src/kernel/cap/capability.c
@@ -48,14 +48,14 @@ static cap_id_t cap_encode(uint16_t slot, uint16_t generation) {
     return (cap_id_t)((generation << CAP_SLOT_BITS) | slot);
 }
 
-static int cap_decode(cap_id_t cap, uint16_t *slot, uint16_t *generation) {
-    uint16_t raw;
+static int cap_decode(cap_id_t cap, uint32_t *slot, uint32_t *generation) {
+    uint32_t raw;
 
     if (cap == CAP_INVALID || cap < 0) {
         return 0;
     }
 
-    raw = (uint16_t)cap;
+    raw = (uint32_t)cap;
     *slot = raw & CAP_SLOT_MASK;
     *generation = raw >> CAP_SLOT_BITS;
     if (*slot >= CAP_MAX_COUNT_VAL || *generation == 0) {
@@ -65,8 +65,8 @@ static int cap_decode(cap_id_t cap, uint16_t *slot, uint16_t *generation) {
 }
 
 static cap_entry_t *cap_get_entry(cap_id_t cap) {
-    uint16_t slot;
-    uint16_t generation;
+    uint32_t slot;
+    uint32_t generation;
 
     if (!cap_decode(cap, &slot, &generation)) {
         return NULL;
